og: Add new route and refactor existing code into admin add and author add.
