+++ src/app/games/admin/add/game.add.original.pug
@@ -1,0 +1,30 @@
+include ../../../template/_mixins
+include ./game.add.mixins.pug
+
+.row.row--fullheight
+	.col-sm-9.col--main
+		h1.h--primary add original game
+		hr
+		p Note that a game is completely decoupled from its releases. Data appearing in here is common to all releases
+			| that are linked to that game, but kept as a separate and independent data set.
+
+		form(role="form")
+
+			+title('1. Enter Title')
+
+		.clearfix
+		h1 2. Game Info
+		hr
+
+		+originalGameInfo()
+		.clearfix
+
+		+mediaUpload('3. Upload Media')
+
+		hr
+		button#game-reset-btn.btn.btn-default.btn-lg(ng-click="vm.reset()") Reset
+		button#game-submit-btn.btn.btn-primary.btn-lg.pull-right(ng-click="vm.submit()", ng-disabled="vm.submitting") Add New Game
+		.clearfix
+
+		p
+			br
